requirements.txt
appdirs==1.4.4
bcrypt==4.0.1
black==22.10.0
cffi==1.15.1
click==8.1.3
distro==1.8.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-prometheus-metrics==1.0.0
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonformatter==0.3.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.10.1
platformdirs==2.5.2
prometheus-client==0.15.0
pycparser==2.21
pyhdb==0.3.4
regex==2022.10.31
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
waitress==2.1.2
Werkzeug==2.2.2
